+++ crates/sui-indexer/src/handlers/mod.rs
@@ -19,7 +19,7 @@ use crate::{
 
 pub mod checkpoint_handler;
 pub mod committer;
-pub mod objects_snapshot_processor;
+pub mod objects_snapshot_handler;
 pub mod pruner;
 pub mod tx_processor;
