+++ crates/sui-indexer/src/handlers/objects_snapshot_handler.rs
